# server_swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/statistics/trending": {
      "get": {
        "operationId": "StatisticsController_findTrending",
        "summary": "",
        "description": "Trending refers to \"popular in the last week/month",
        "parameters": [
          {
            "name": "sourceType",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "game",
                "review",
                "activity",
                "collection"
              ],
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsPaginatedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "statistics"
        ]
      }
    },
    "/v1/statistics": {
      "get": {
        "operationId": "StatisticsController_findOne",
        "parameters": [
          {
            "name": "sourceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceType",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "game",
                "review",
                "activity",
                "collection"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsEntityDto"
                }
              }
            }
          }
        },
        "tags": [
          "statistics"
        ]
      }
    },
...
  },
  "info": {
    "title": "GameNode API",
    "description": "API docs for the videogame catalog system GameNode. <br><br>Built with love by the GameNode team.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ProfileAvatar": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "mimetype": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "filename": {
            "type": "string"
          },
          "encoding": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Multer's File's destination + filename"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mimetype",
          "extension",
          "size",
          "filename",
          "encoding",
          "path",
          "profile",
          "createdAt",
          "updatedAt"
        ]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "Shareable string ID\n\nSame as SuperTokens' userId."
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "$ref": "#/components/schemas/ProfileAvatar"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "username",
          "avatar",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "statistics": {
            "$ref": "#/components/schemas/Statistics"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "statistics",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserLike": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "statistics": {
            "$ref": "#/components/schemas/Statistics"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "profile",
          "statistics",
          "createdAt",
          "updatedAt"
        ]
      },
      "Statistics": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sourceId": {
            "type": "string",
            "description": "Always stored as string. May sometimes refer to a number ID."
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "game",
              "review",
              "activity",
              "collection"
            ]
          },
          "viewsCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          },
          "views": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserView"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserLike"
            }
          }
        },
        "required": [
          "id",
          "sourceId",
          "sourceType",
          "viewsCount",
          "likesCount",
          "views",
          "likes"
        ]
      },
      "StatisticsPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Statistics"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "StatisticsEntityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "sourceId": {
            "type": "string",
            "description": "Always stored as string. May sometimes refer to a number ID."
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "game",
              "review",
              "activity",
              "collection"
            ]
          },
          "viewsCount": {
            "type": "number"
          },
          "likesCount": {
            "type": "number"
          },
          "isLiked": {
            "type": "boolean"
          },
          "isViewed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "sourceId",
          "sourceType",
          "viewsCount",
          "likesCount",
          "isLiked",
          "isViewed"
        ]
      },
...
    }
  }
}